Update usage of `ConfigurationAdapter` in Velocity
On `dev/dev`, `ConfigurationAdapter#saveDefaultConfig` could be used to save the default settings for a platform to a file. However, since Spigot and Paper have different config options, `DefaultBukkitConfig` was updated to handle both Paper and Spigot configs. Therefore, `BukkitConfigurationAdapter#saveDefaultConfig` did not have a single default config, so the signature of `saveDefaultConfig` was updated to input the instance of `DefaultConfig` that should be written to the file.

This commit fixes a mistake where Velocity was still trying to use the `saveDefaultConfig` method with only two arguments. Since `saveDefaultConfig` no longer always uses *the* default config for the platform, I also renamed the method to `saveConfig`.

Something to think about in code review: `DefaultBukkitConfig` is in `commandapi-bukkit`, but it contains the Paper-specific field `SHOULD_HOOK_PAPER_RELOAD`. This field does not need to exist on Spigot, so should it be moved into a Paper-specific subclass of `DefaultBukkitConfig`, similar to what happens with `CommandAPIBukkitConfig` and `CommandAPIPaperConfig`?
Also, do we need 4 `CommandAPIMain` classes? Can we put them into a common module like before? Perhaps the command conversion logic can be extracted into the Bukkit module, and the config stuff can be extracted into a `spigot/paper-plugin-common`