Improve DynamicMultiLiteralArgument on Spigot
Before this commit, since we can't modify the CommandTree on a per-player basis, DynamicMultiLiteralArguments weren't dynamic. Now, players are told this is a StringArgument with suggestions, which means the argument doesn't act like a literal (invalid input appears to be accepted for the client). I think it makes more sense this way since the suggestions are correct and valid input will be accepted.